# Upgrade go storage client version