### Description
Upgrade go storage client library version with `go get cloud.google.com/go/storage@latest`
Used storage.WithJSONReads() to read with JSON api.

### Link to the issue in case of a bug fix.
NA

### Testing details
1. Manual - NA
2. Unit tests - NA
3. Integration tests - Executing kokoro perf tests.
+--------+------------+--------------+------------+--------------+--------------+
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
+--------+------------+--------------+------------+--------------+--------------+
| Master |  0.25MiB   | 446.98MiB/s  | 1.84MiB/s  |  31.0MiB/s   |  2.55MiB/s   |
|   PR   |  0.25MiB   | 445.74MiB/s  | 1.98MiB/s  |  31.82MiB/s  |   1.9MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 4302.08MiB/s | 79.21MiB/s | 1096.73MiB/s |  80.15MiB/s  |
|   PR   | 48.828MiB  | 4321.74MiB/s | 80.3MiB/s  | 1162.31MiB/s |  80.13MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 4419.47MiB/s | 25.34MiB/s | 684.16MiB/s  |  33.7MiB/s   |
|   PR   | 976.562MiB | 4330.65MiB/s | 31.88MiB/s | 786.95MiB/s  |  32.91MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
+--------+------------+--------------+------------+--------------+--------------+